# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion des Comptes API",
        "description": "API pour la gestion des comptes, incluant toutes les opérations CRUD, archivage, blocage et déblocage.",
        "contact": {
            "email": "support@gestionapi.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/archives/semaine/{semaineId}": {
            "get": {
                "tags": [
                    "Archives"
                ],
                "summary": "Lister les transactions archivées d'une semaine",
                "operationId": "b0c268f2873efcb405bdd78467cb4376",
                "parameters": [
                    {
                        "name": "semaineId",
                        "in": "path",
                        "description": "Identifiant de la semaine (ex: 2025-44)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des transactions archivées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/dashboard/global": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Dashboard global (admin)",
                "operationId": "110b8172fb1000e422826ce05427eca7",
                "responses": {
                    "200": {
                        "description": "Statistiques globales",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/dashboard/client/{clientId}": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Dashboard d'un client",
                "operationId": "1b0b386f84c00ce53b69747947278c6b",
                "parameters": [
                    {
                        "name": "clientId",
                        "in": "path",
                        "description": "ID du client",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistiques du client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Lister toutes les transactions",
                "operationId": "64732d889d5281eec388e75a8febda6e",
                "responses": {
                    "200": {
                        "description": "Liste des transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/transactions/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Récupérer une transaction par id",
                "operationId": "dba4903b2c796853c7d2b4637832a0fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la transaction",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Annuler/supprimer une transaction (admin)",
                "operationId": "3782d1e6138f3fe6dc938731adc8f74c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la transaction",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}